Vodafone/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
reportlab==4.0.0
requests==2.31.0
feedparser==6.0.10

# SWHI Dependencies
dnspython==2.4.2
python-whois==0.8.0

# SWHI ML Dependencies (optional)
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
